[DSE] Add first version of MemorySSA-backed DSE.
Summary:
This patch adds a first version of a MemorySSA based DSE. It is missing
a lot of features, which will get added as follow-ups, to help to keep
the review manageable.

For each MemoryDef, we walk the MemorySSA def-use chains to try to find
a overwriting MemoryDef. During the walk, we bail out if we hit either:
 * A MemoryPhi (restriction will be lifted later).
 * An aliasing MemoryAccess.
 * Multiple MemoryDefs (which indicate a split in control flow).

The patch uses a very simple approach to guarantee that no throwing
instructions are between 2 stores: We only allow accesses to stack
objects, access that are in the same basic block if the block does not
contain any throwing instructions or accesses in functions that do
not contain any throwing instructions. This will get lifted later.

Besides adding support for the missing cases, there is plenty of additonal
potential for improvements as follow-up work, e.g. the way we visit stores
(could be just a traversal of the MemorySSA, rather than collecting them
up-front), using the alias information discovered during walking to optimize
the MemorySSA.

This is based on D40480 by Dave Green.

Reviewers: dmgreen, rnk, efriedma, bryant, asbirlea

Subscribers: Tyker, tschuett, merge_guards_bot, Prazek, hiraditya, george.burgess.iv, jfb, llvm-commits

Tags: #llvm

Differential Revision: https://reviews.llvm.org/D72146